refactor(node): Extract `filter_logs` into a shared function
